+++ account/views.py
@@ -1,6 +1,6 @@
 from django.contrib.auth import get_user_model
 from rest_framework import viewsets
-from .Serializers import UserSerializer
+from .serializers import UserSerializer
 User = get_user_model()
